+++ src/xcode/ENA/ENA/Source/HTTPClient/Service/Services/ServiceType.swift
@@ -8,4 +8,5 @@ enum ServiceType {
 	case caching
 	case wifiOnly
 	case retrying
+	case dynamicPinning
 }
