+++ src/xcode/ENA/ENA/Source/HTTPClient/Service/Services/StandardRestService.swift
@@ -2,8 +2,6 @@
 // 🦠 Corona-Warn-App
 //
 
-import Foundation
-
 /**
 Specific implementation of a service who is doing nothing special, so it is the default one.
 It uses the coronaWarnSessionConfiguration.
